Block hash validation test moved to BlockValidationLogicSpec
